# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "AnswerQuestionRequestDto": {
        "properties": {
          "selectedOptionId": {
            "maxLength": 191,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "selectedOptionId"
        ],
        "type": "object"
      },
      "AnswerResultDto": {
        "properties": {
          "balance": {
            "minimum": 0,
            "type": "number"
          },
          "correct": {
            "type": "boolean"
          },
          "correctOptionId": {
            "type": "string"
          },
          "errorCount": {
            "minimum": 0,
            "type": "number"
          },
          "explanation": {
            "type": "string"
          },
          "pointsAwarded": {
            "minimum": 0,
            "type": "number"
          },
          "selectedOptionId": {
            "type": "string"
          }
        },
        "required": [
          "correct",
          "selectedOptionId",
          "correctOptionId",
          "explanation",
          "errorCount",
          "pointsAwarded",
          "balance"
        ],
        "type": "object"
      },
      "ApiErrorDto": {
        "properties": {
          "code": {
            "example": "VALIDATION_FAILED",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "type": "object"
          },
          "message": {
            "example": "请求参数验证失败",
            "type": "string"
          },
          "requestId": {
            "example": "7da2aa93-ef82-45c8-9df0-84232e6a5b13",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "requestId",
          "details"
        ],
        "type": "object"
      },
...
      "LearnerQuestionDto": {
        "properties": {
          "basePoints": {
            "type": "number"
          },
          "id": {
            "type": "string"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/LearnerQuestionOptionDto"
            },
            "type": "array"
          },
          "stem": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "stem",
          "basePoints",
          "options"
        ],
        "type": "object"
      },
      "LearnerQuestionOptionDto": {
        "properties": {
          "content": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "position": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "label",
          "content",
          "position"
        ],
        "type": "object"
      },
...
      "PageMetaDto": {
        "properties": {
          "page": {
            "example": 1,
            "minimum": 1,
            "type": "number"
          },
          "pageSize": {
            "example": 20,
            "maximum": 100,
            "minimum": 1,
            "type": "number"
          },
          "total": {
            "example": 42,
            "minimum": 0,
            "type": "number"
          },
          "totalPages": {
            "example": 3,
            "minimum": 0,
            "type": "number"
          }
        },
        "required": [
          "page",
          "pageSize",
          "total",
          "totalPages"
        ],
        "type": "object"
      },
...
      "PracticeSummaryDto": {
        "properties": {
          "activeTotal": {
            "minimum": 0,
            "type": "number"
          },
          "balance": {
            "minimum": 0,
            "type": "number"
          },
          "firstAnsweredCount": {
            "minimum": 0,
            "type": "number"
          },
          "masteredWrongCount": {
            "minimum": 0,
            "type": "number"
          },
          "pendingWrongCount": {
            "minimum": 0,
            "type": "number"
          },
          "unansweredCount": {
            "minimum": 0,
            "type": "number"
          }
        },
        "required": [
          "activeTotal",
          "firstAnsweredCount",
          "unansweredCount",
          "pendingWrongCount",
          "masteredWrongCount",
          "balance"
        ],
        "type": "object"
      },
...
      "WrongQuestionItemDto": {
        "properties": {
          "errorCount": {
            "minimum": 1,
            "type": "number"
          },
          "firstAnsweredAt": {
            "format": "date-time",
            "type": "string"
          },
          "masteredAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "question": {
            "$ref": "#/components/schemas/LearnerQuestionDto"
          }
        },
        "required": [
          "question",
          "errorCount",
          "firstAnsweredAt",
          "masteredAt"
        ],
        "type": "object"
      },
      "WrongQuestionListResponseDto": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/WrongQuestionItemDto"
            },
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "Android 与其他无 Cookie 客户端使用的访问令牌",
        "scheme": "bearer",
        "type": "http"
      },
      "cookieAuth": {
        "description": "Web 客户端访问令牌 Cookie",
        "in": "cookie",
        "name": "pq_access",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "英语答题、积分、商品兑换与订单管理开放接口",
    "title": "Point Quest API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/practice/questions/{questionId}/answer": {
      "post": {
        "operationId": "practiceAnswerQuestion",
        "parameters": [
          {
            "in": "path",
            "name": "questionId",
            "required": true,
            "schema": {
              "maxLength": 191,
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "description": "使用 Cookie 身份认证执行写操作时必填；Bearer 模式勿填",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "description": "同一用户内唯一，重试同一请求时必须复用",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerQuestionRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerResultDto"
                }
              }
            },
            "description": "成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求参数验证失败"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "身份认证失败"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "权限不足或 CSRF 校验失败"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "资源不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "状态、幂等或并发冲突"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求体或上传文件过大"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "服务器内部错误"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "summary": "提交题目首次答案",
        "tags": [
          "学生端-练习"
        ]
      }
    },
    "/api/v1/practice/random": {
      "get": {
        "operationId": "practiceGetRandomQuestion",
        "parameters": [
          {
            "description": "本次客户端会话需排除的题目 ID，使用逗号分隔，最多 50 个",
            "in": "query",
            "name": "excludeIds",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearnerQuestionDto"
                }
              }
            },
            "description": "成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求参数验证失败"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "身份认证失败"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "权限不足或 CSRF 校验失败"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "资源不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "状态、幂等或并发冲突"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求体或上传文件过大"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "服务器内部错误"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "summary": "随机获取一题未答题目",
        "tags": [
          "学生端-练习"
        ]
      }
    },
    "/api/v1/practice/summary": {
      "get": {
        "operationId": "practiceGetSummary",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeSummaryDto"
                }
              }
            },
            "description": "成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求参数验证失败"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "身份认证失败"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "权限不足或 CSRF 校验失败"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "资源不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "状态、幂等或并发冲突"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求体或上传文件过大"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "服务器内部错误"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "summary": "获取练习统计摘要",
        "tags": [
          "学生端-练习"
        ]
      }
    },
    "/api/v1/practice/wrong-questions": {
      "get": {
        "operationId": "practiceListWrongQuestions",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 100000,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrongQuestionListResponseDto"
                }
              }
            },
            "description": "成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求参数验证失败"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "身份认证失败"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "权限不足或 CSRF 校验失败"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "资源不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "状态、幂等或并发冲突"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求体或上传文件过大"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "服务器内部错误"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "summary": "分页查询待练错题",
        "tags": [
          "学生端-练习"
        ]
      }
    },
    "/api/v1/practice/wrong-questions/{questionId}/answer": {
      "post": {
        "operationId": "practiceRetryWrongQuestion",
        "parameters": [
          {
            "in": "path",
            "name": "questionId",
            "required": true,
            "schema": {
              "maxLength": 191,
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "description": "使用 Cookie 身份认证执行写操作时必填；Bearer 模式勿填",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "description": "同一用户内唯一，重试同一请求时必须复用",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerQuestionRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerResultDto"
                }
              }
            },
            "description": "成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求参数验证失败"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "身份认证失败"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "权限不足或 CSRF 校验失败"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "资源不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "状态、幂等或并发冲突"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求体或上传文件过大"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "服务器内部错误"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "summary": "提交错题重练答案",
        "tags": [
          "学生端-练习"
        ]
      }
    },
...
  },
  "servers": [],
...
}